pyproject.toml
[project]
name = "testedjango"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-ninja>=1.4.5",
]
